# Add Benchee benchmarks for Plugs