--- a/benchmarks/support/tracer.ex
+++ b/benchmarks/support/tracer.ex
@@ -1,0 +1,4 @@
+defmodule Benchmark.Tracer do
+  @moduledoc false
+  use Spandex.Tracer, otp_app: :benchmark
+end