This is an initial stab at adding [Benchee](https://github.com/PragTob/benchee) benchmarks for the `Spandex.Plug.*` modules to get a feel for how much performance impact there is when adding Spandex to a service. It is important to note that since the `BaselineRouter` doesn't really do anything, the `x slower` metrics are meaningless, so we should focus on the raw per-request timing increases due to adding tracing (in terms of how many microseconds is tracing adding to each request we process).

I wanted to get this in place so that I can work on an implementation for trace sampling (hence the stubbed-in `inputs` list) and validate that it alleviates the performance overhead of tracing, at least on average, when only sampling a small fraction of requests.

Feedback welcome!

Some initial results from my laptop (with a million other things running, so take it with a grain of salt):

```
Operating System: macOS"
CPU Information: Intel(R) Core(TM) i7-7920HQ CPU @ 3.10GHz
Number of Available Cores: 8
Available memory: 16 GB
Elixir 1.6.6
Erlang 20.3.8.1

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 10 s
memory time: 2 s
parallel: 4
inputs: no_sampling
Estimated total run time: 42 s


Benchmarking addcontext with input no_sampling...
Benchmarking baseline with input no_sampling...
Benchmarking start_finish with input no_sampling...

##### With input no_sampling #####
Name                   ips        average  deviation         median         99th %
baseline           49.04 K       20.39 μs    ±82.41%          18 μs          44 μs
start_finish        5.62 K      178.03 μs    ±45.97%         168 μs         427 μs
addcontext          4.74 K      210.96 μs    ±73.94%         185 μs      771.44 μs

Comparison:
baseline           49.04 K
start_finish        5.62 K - 8.73x slower
addcontext          4.74 K - 10.35x slower

Memory usage statistics:

Name                 average  deviation         median         99th %
baseline             3.81 KB     ±0.00%        3.81 KB        3.81 KB
start_finish        24.76 KB     ±0.02%       24.77 KB       24.77 KB
addcontext          33.67 KB     ±0.02%       33.67 KB       33.67 KB

Comparison:
baseline             3.81 KB
start_finish        24.77 KB - 6.50x memory usage
addcontext          33.67 KB - 8.83x memory usage
```