@zachdaniel 2018-08-17
This looks good to me. Let me know when you are ready and I will merge.