@GregMefford 2018-08-17
I think this is ready to merge when you are. I'll send the sampling thing as a separate PR when I get it figured out. The only issue with this one is that it's currently tangled with the Datadog sender API because it wasn't obvious to be how to use the test ones you made, and actually it kind of makes sense to test the sender when doing a benchmark since it's a thing that happens in production use-cases. Maybe this benchmark should only test the `Spandex` code, though, and we would separately benchmark the `SpandexDatadog` code? I'll need to think more about that, but I think this PR is fine for now.