@zachdaniel 2018-08-17
I think I might want to merge the adapter split first, and then merge this. There is a test adapter that is platform agnostic that could be used for these benchmarks in that new branch.