@GregMefford 2018-08-18
Yep, I agree that’s a better choice.

On Fri, Aug 17, 2018 at 6:02 PM Zach Daniel <notifications@github.com>
wrote:

> I think I might want to merge the adapter split first, and then merge
> this. There is a test adapter that is platform agnostic that could be used
> for these benchmarks in that new branch.
>
> —
> You are receiving this because you authored the thread.
>
>
> Reply to this email directly, view it on GitHub
> <https://github.com/zachdaniel/spandex/pull/59#issuecomment-413999104>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAEPW14O9VDyvz_e9rHzWdHhW8rN15Z_ks5uRz1egaJpZM4WAgoi>
> .
>