@zachdaniel 2018-08-18
Alright, the other branch is merged. Not going to cut a release just yet though, want to let it mellow and test it out a bit.