@GregMefford 2018-08-18
Cool. I will work on rebasing this to use the test adapter.

On Fri, Aug 17, 2018 at 10:51 PM Zach Daniel <notifications@github.com>
wrote:

> Alright, the other branch is merged. Not going to cut a release just yet
> though, want to let it mellow and test it out a bit.
>
> —
> You are receiving this because you authored the thread.
>
>
> Reply to this email directly, view it on GitHub
> <https://github.com/zachdaniel/spandex/pull/59#issuecomment-414027490>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAEPWwo5u8iT5ey0iIW7YhaiO3BzCYO-ks5uR4E8gaJpZM4WAgoi>
> .
>